Fix datatypes (#10)
* use uint64 type for stakedtokens as int32 is too small


* Add changes to correctly handle huge numbers of staked tokens where uint64 is not enough (e.g. injective).


---------

+++ internal/packages/consensus/uptime/collector/collector.go
@@ -240,7 +240,7 @@ func loop(exporter *common.Exporter, p common.Packager) {
 						common.ProposerAddressLabel:  item.ProposerAddress,
 						common.MonikerLabel:          item.Moniker,
 					}).
-					Set(float64(item.StakedTokens))
+					Set(item.StakedTokens)
 
 				validatorCommissionMetric.
 					With(prometheus.Labels{
@@ -279,7 +279,7 @@ func loop(exporter *common.Exporter, p common.Packager) {
 							common.ProposerAddressLabel:  item.ProposerAddress,
 							common.MonikerLabel:          item.Moniker,
 						}).
-						Set(float64(item.StakedTokens))
+						Set(item.StakedTokens)
 					validatorCommissionMetric.
 						With(prometheus.Labels{
 							common.ValidatorAddressLabel: item.ValidatorOperatorAddress,
